in-class comments added
