+++ nodetest2/app/index.html
@@ -1,13 +1,13 @@
 <!doctype html>
 <html class="no-js" lang="EN" data-ng-app="userApp">
   <head>
-    <meta charset="utf-8">
-    <title></title>
-    <link rel="stylesheet" href="css/style.css">
+    <title>User Management</title>
+    <link rel="stylesheet" href="css/index.css">
   </head>
-
   <body>
+
     <main data-ng-controller="userController">
+
       <ul id="addUserChunk">
         <li>
           <h2 id="addUser">Add User</h2>
@@ -92,7 +92,7 @@
 
     </main>
 
-    <script src="bundle.js"></script>
+    <script src="js/bundle.js"></script>
   </body>
 </html>
