+++ nodetest2/app/js/client.js
@@ -1,0 +1,7 @@
+'use strict';
+
+require('angular/angular');
+
+var userApp = angular.module('userApp', []);
+
+require("./users/usersController")(userApp);
