--- nodetest2/data/mongod.lock
@@ -1,1 +1,0 @@
-533
