+++ nodetest2/public/index.html
@@ -1,1 +1,1 @@
-<!doctype html><html class=no-js lang=EN data-ng-app=userApp><head><meta charset=utf-8><title></title><link rel=stylesheet href=css/style.css></head><body><main data-ng-controller=userController><ul id=addUserChunk><li><h2 id=addUser>Add User</h2></li><ul><form data-ng-submit=submitForm(user)><fieldset><li><input id=inputUserName type=text ,="" placeholder=Username data-ng-model=user.userName></li><li><input id=inputUserEmail type=text ,="" placeholder=Email data-ng-model=user.userEmail></li><li><input id=inputUserFullname type=text ,="" placeholder="Full Name" data-ng-model=user.userFullName></li><li><input id=inputUserAge type=text ,="" placeholder=Age data-ng-model=user.userAge></li><li><input id=inputUserLocation type=text ,="" placeholder=Location data-ng-model=user.userLocation></li><li><input id=inputUserGender type=text ,="" placeholder=Gender data-ng-model=user.userGender></li><li><button id=btnAddUser type=submit>Add User</button></li></fieldset></form></ul></ul><ul id=#updateInfoChunk><li><h2 id=updateInfo>Update Info</h2></li><ul><fieldset><li><input id=updateUserName type=text ,="" placeholder=Username data-ng-model=user.userName></li><li><input id=updateUserEmail type=text ,="" placeholder=Email data-ng-model=user.userEmail></li><li><input id=updateUserFullname type=text ,="" placeholder="Full Name" data-ng-model=user.userFullName></li><li><input id=updateUserAge type=text ,="" placeholder=Age data-ng-model=user.userAge></li><li><input id=updateUserLocation type=text ,="" placeholder=Location data-ng-model=user.userLocation></li><li><input id=updateUserGender type=text ,="" placeholder=Gender data-ng-model=user.userGender></li><li><button id=btnUpdateInfo data-ng-click=edit(user)>Update User</button></li><li><button id=btnCancelUpdateInfo>Cancel</button></li></fieldset></ul></ul><h2>User List</h2><ul><li data-ng-repeat="user in users"><h3>{{user.name}}</h3><ul><li>Username: {{user.params.userName}}</li><li>Email: {{user.params.userEmail}}</li><li>Full Name: {{user.params.userFullName}}</li><li>Age: {{user.params.userAge}}</li><li>Location: {{user.params.userLocation}}</li><li>Gender: {{user.params.userGender}}</li></ul></li></ul></main><script src=bundle.js></script></body></html>
+<!doctype html><html class=no-js lang=EN data-ng-app=userApp><head><title>User Management</title><link rel=stylesheet href=css/index.css></head><body><main data-ng-controller=userController><ul id=addUserChunk><li><h2 id=addUser>Add User</h2></li><ul><form data-ng-submit=submitForm(user)><fieldset><li><input id=inputUserName type=text ,="" placeholder=Username data-ng-model=user.userName></li><li><input id=inputUserEmail type=text ,="" placeholder=Email data-ng-model=user.userEmail></li><li><input id=inputUserFullname type=text ,="" placeholder="Full Name" data-ng-model=user.userFullName></li><li><input id=inputUserAge type=text ,="" placeholder=Age data-ng-model=user.userAge></li><li><input id=inputUserLocation type=text ,="" placeholder=Location data-ng-model=user.userLocation></li><li><input id=inputUserGender type=text ,="" placeholder=Gender data-ng-model=user.userGender></li><li><button id=btnAddUser type=submit>Add User</button></li></fieldset></form></ul></ul><ul id=#updateInfoChunk><li><h2 id=updateInfo>Update Info</h2></li><ul><fieldset><li><input id=updateUserName type=text ,="" placeholder=Username data-ng-model=user.userName></li><li><input id=updateUserEmail type=text ,="" placeholder=Email data-ng-model=user.userEmail></li><li><input id=updateUserFullname type=text ,="" placeholder="Full Name" data-ng-model=user.userFullName></li><li><input id=updateUserAge type=text ,="" placeholder=Age data-ng-model=user.userAge></li><li><input id=updateUserLocation type=text ,="" placeholder=Location data-ng-model=user.userLocation></li><li><input id=updateUserGender type=text ,="" placeholder=Gender data-ng-model=user.userGender></li><li><button id=btnUpdateInfo data-ng-click=edit(user)>Update User</button></li><li><button id=btnCancelUpdateInfo>Cancel</button></li></fieldset></ul></ul><h2>User List</h2><ul><li data-ng-repeat="user in users"><h3>{{user.name}}</h3><ul><li>Username: {{user.params.userName}}</li><li>Email: {{user.params.userEmail}}</li><li>Full Name: {{user.params.userFullName}}</li><li>Age: {{user.params.userAge}}</li><li>Location: {{user.params.userLocation}}</li><li>Gender: {{user.params.userGender}}</li></ul></li></ul></main><script src=js/bundle.js></script></body></html>
